Fix regression causing off-by-one pager height on truncated autosuggestion
Commit c54131c8c5 (Make at least one character of the autosuggestion
always shown, 2025-01-26) draws autosuggetion until exactly the end
of the line, possibly truncated by an ellipsis.  The line is "barely"
soft-wrapped; no characters or cursors are on the next line.

This means that the pager can start there (on the next line).
The calculation for available pager height fails to recognize that.
Fix that.